RRT smoothing and hysteresis. Fixed timestamp bugs.
git-svn-id: https://svn.robojackets.org/robocup/software/trunk@834 7ccbd594-6f30-0410-bcae-e071bb1c4cdc

+++ Vision/vision/Process.cpp
@@ -139,5 +139,6 @@ uint64_t Process::timestamp() const
 	struct timeval time;
 	gettimeofday(&time, 0);
 	
-	return time.tv_sec * 1000000 + time.tv_usec;
+	return (uint64_t)time.tv_sec * 1000000 + (uint64_t)time.tv_usec;
 }
+
